Revert "deps+perf(server): upgraded scodec"
This reverts commit 6b74422a1b82769d93907a7e320c30b816703961.

+++ actor-server/actor-models/src/main/scala/im/actor/server/mtproto/protocol/ProtoMessage.scala
@@ -17,7 +17,7 @@ sealed trait IncomingProtoMessage
 sealed trait OutgoingProtoMessage
 
 sealed trait ResendableProtoMessage {
-  def bodySize: Long
+  def bodySize: Int
 }
 
 @SerialVersionUID(1L)
